+++ libs/x86/arch/mmu.h
@@ -55,7 +55,7 @@ struct mmu_directory
 
 };
 
-void mmu_map(struct mmu_directory *directory, struct mmu_table *table, unsigned int paddress, unsigned int vaddress, unsigned int size, unsigned int tflags, unsigned int pflags);
+void mmu_map(struct mmu_directory *directory, struct mmu_table *table, unsigned long paddress, unsigned long vaddress, unsigned int size, unsigned int tflags, unsigned int pflags);
 void mmu_enable();
 void mmu_reload();
 void mmu_load(struct mmu_directory *directory);
