+++ modules/arch/x86/i915/driver.c
@@ -9,7 +9,7 @@
 static unsigned int read(unsigned int reg)
 {
 
-    volatile unsigned int *value = (volatile unsigned int *)reg;
+    volatile unsigned int *value = (volatile unsigned int *)(unsigned long)reg;
 
     return *value;
 
@@ -18,7 +18,7 @@ static unsigned int read(unsigned int reg)
 static void write(unsigned int reg, unsigned int val)
 {
 
-    volatile unsigned int *value = (volatile unsigned int *)reg;
+    volatile unsigned int *value = (volatile unsigned int *)(unsigned long)reg;
 
     *value = val;
