+++ modules/system/protocol.c
@@ -10,16 +10,16 @@ static unsigned int root(struct vfs_backend *backend)
 
     backend->read(backend, 0, sizeof (struct system_header), &header);
 
-    return (unsigned int)header.root;
+    return (unsigned long)header.root;
 
 }
 
 static unsigned int parent(struct vfs_backend *backend, unsigned int id)
 {
 
-    struct system_node *node = (struct system_node *)id;
+    struct system_node *node = (struct system_node *)(unsigned long)id;
 
-    return (unsigned int)node->parent;
+    return (unsigned long)node->parent;
 
 }
 
@@ -37,7 +37,7 @@ static unsigned int match(struct vfs_backend *backend)
 static unsigned int open(struct vfs_backend *backend, unsigned int id)
 {
 
-    struct system_node *node = (struct system_node *)id;
+    struct system_node *node = (struct system_node *)(unsigned long)id;
 
     return node->open(node);
 
@@ -46,7 +46,7 @@ static unsigned int open(struct vfs_backend *backend, unsigned int id)
 static unsigned int close(struct vfs_backend *backend, unsigned int id)
 {
 
-    struct system_node *node = (struct system_node *)id;
+    struct system_node *node = (struct system_node *)(unsigned long)id;
 
     return node->close(node);
 
@@ -55,7 +55,7 @@ static unsigned int close(struct vfs_backend *backend, unsigned int id)
 static unsigned int read(struct vfs_backend *backend, unsigned int id, unsigned int offset, unsigned int count, void *buffer)
 {
 
-    struct system_node *node = (struct system_node *)id;
+    struct system_node *node = (struct system_node *)(unsigned long)id;
 
     return node->read(node, offset, count, buffer);
 
@@ -64,7 +64,7 @@ static unsigned int read(struct vfs_backend *backend, unsigned int id, unsigned 
 static unsigned int walk(struct vfs_backend *backend, unsigned int id, unsigned int count, const char *path)
 {
 
-    struct system_node *node = (struct system_node *)id;
+    struct system_node *node = (struct system_node *)(unsigned long)id;
 
     return node->walk(node, count, path);
 
@@ -73,7 +73,7 @@ static unsigned int walk(struct vfs_backend *backend, unsigned int id, unsigned 
 static unsigned int write(struct vfs_backend *backend, unsigned int id, unsigned int offset, unsigned int count, void *buffer)
 {
 
-    struct system_node *node = (struct system_node *)id;
+    struct system_node *node = (struct system_node *)(unsigned long)id;
 
     return node->write(node, offset, count, buffer);
